requirements.txt
# These requirements are for local development and testing of the module

# python packaging
twine==3.1.1

# django stuff
Django==3.1.9
djangorestframework==3.12.4

# factories
factory-boy==3.2.0

# tests
pytest==5.4.1
pytest-django==3.9.0
pytest-pythonpath==0.7.3
flake8==3.7.9
